+++ test/issues.ts
@@ -26,7 +26,7 @@ describe('Issues', function () {
                 prodId: '//superman-industries.com//ical-generator//EN',
                 timezone: 'Brazil/East',
                 events: [{
-                    start: moment('Sun May 01 2016 00:00:00 GMT-3'),
+                    start: moment('2016-05-01T03:00:00.000Z'),
                     summary: 'Example Event',
                     allDay: true
                 }]
@@ -42,8 +42,8 @@ describe('Issues', function () {
             const calendar = ical({
                 prodId: '//superman-industries.com//ical-generator//EN',
                 events: [{
-                    start: moment('Sun May 01 2016 00:00:00 GMT+0200 (CEST)'),
-                    end: moment('Sun May 01 2016 02:00:00 GMT+0200 (CEST)'),
+                    start: moment('2016-04-30T22:00:00.000Z'),
+                    end: moment('2016-05-01T00:00:00.000Z'),
                     summary: 'Example Event',
                     allDay: true,
                     repeating: {
@@ -60,12 +60,12 @@ describe('Issues', function () {
             assert.ok(str.indexOf('RRULE:FREQ=MONTHLY;COUNT=3;INTERVAL=1;BYDAY=SU;BYSETPOS=3') > -1);
         });
 
-        it('should work with repeating bySetPos by taking the first elemnt of the byDay array', function () {
+        it('should work with repeating bySetPos by taking all elements of the byDay array', function () {
             const calendar = ical({
                 prodId: '//superman-industries.com//ical-generator//EN',
                 events: [{
-                    start: moment('Sun May 01 2016 00:00:00 GMT+0200 (CEST)'),
-                    end: moment('Sun May 01 2016 02:00:00 GMT+0200 (CEST)'),
+                    start: moment('2016-04-30T22:00:00.000Z'),
+                    end: moment('2016-05-01T00:00:00.000Z'),
                     summary: 'Example Event',
                     allDay: true,
                     repeating: {
@@ -79,7 +79,7 @@ describe('Issues', function () {
             });
 
             const str = calendar.toString();
-            assert.ok(str.indexOf('RRULE:FREQ=MONTHLY;COUNT=3;INTERVAL=1;BYDAY=MO;BYSETPOS=3') > -1);
+            assert.ok(str.indexOf('RRULE:FREQ=MONTHLY;COUNT=3;INTERVAL=1;BYDAY=MO,FR;BYSETPOS=3') > -1);
         });
     });
 
@@ -209,5 +209,10 @@ describe('Issues', function () {
                 'END:VCALENDAR'
             ].join('\r\n'));
         });
+    });
+
+    describe('Issue #442', function () {
+        it('should generate floating repeat until value if event is a floating event');
+        it('should generate floating repeat exclusion dates if event is a floating event');
     });
 });